pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cytoolz>=0.12.3",
    "loguru>=0.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.basedpyright]
include = ["**"]
typeCheckingMode = "basic"
venv = "."
venvPath = ".venv"

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
# extra-index-url = []
